+++ books-service/src/test/java/mts/homework/bookService/controllers/TagsControllerTest.java
@@ -1,4 +1,4 @@
-package mts.homework.bookService.controllers.tags;
+package mts.homework.bookService.controllers;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
@@ -20,11 +20,13 @@ import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 import org.testcontainers.junit.jupiter.Testcontainers;
 
+import java.util.Map;
+
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 @Transactional(propagation = Propagation.NOT_SUPPORTED)
 @AutoConfigureTestDatabase(replace = AutoConfigureTestDatabase.Replace.NONE)
 @Testcontainers
-public class TagsControllerCreateTagTest extends DatabaseSuite {
+public class TagsControllerTest extends DatabaseSuite {
   @Autowired private JpaTagsRepository jpaTagsRepository;
 
   @Autowired private TestRestTemplate http;
@@ -55,5 +57,31 @@ public class TagsControllerCreateTagTest extends DatabaseSuite {
 
     var result = http.postForEntity("/api/tags", requestEntity, TagApiEntity.class);
     assertTrue(result.getStatusCode().is4xxClientError());
+  }
+
+  @Test
+  public void deleteTag() {
+    http.delete("/api/tags/{id}", Map.of("id", testTag.getId()));
+
+    assertEquals(0, jpaTagsRepository.findAll().size());
+  }
+
+  @Test
+  public void testTagFind() {
+    var responseResult =
+        http.getForEntity("/api/tags/{id}", TagApiEntity.class, Map.of("id", testTag.getId()));
+
+    var body = TestHelper.assert2xxAndGetBody(responseResult);
+
+    assertEquals(testTag.getId(), body.id());
+    assertEquals(testTag.getName(), body.name());
+  }
+
+  @Test
+  public void testTagFindNotFound() {
+    var responseResult =
+        http.getForEntity("/api/tags/{id}", TagApiEntity.class, Map.of("id", testTag.getId() + 1));
+
+    assertTrue(responseResult.getStatusCode().is4xxClientError());
   }
 }
