+++ books-service/src/test/java/mts/homework/bookService/services/BooksServiceTest.java
@@ -1,4 +1,4 @@
-package mts.homework.bookService.services.books;
+package mts.homework.bookService.services;
 
 import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.ArgumentMatchers.any;
@@ -6,13 +6,13 @@ import static org.mockito.Mockito.when;
 
 import mts.homework.bookService.DatabaseSuite;
 import mts.homework.bookService.data.entities.Author;
+import mts.homework.bookService.data.entities.Book;
+import mts.homework.bookService.data.entities.Tag;
 import mts.homework.bookService.data.repositories.DbBooksRepository;
 import mts.homework.bookService.data.repositories.jpa.JpaAuthorsRepository;
 import mts.homework.bookService.data.repositories.jpa.JpaBooksRepository;
+import mts.homework.bookService.data.repositories.jpa.JpaTagsRepository;
 import mts.homework.bookService.exceptions.InvalidBookDataException;
-import mts.homework.bookService.services.AuthorsRegistryServiceGatewayBase;
-import mts.homework.bookService.services.BookCreationInfo;
-import mts.homework.bookService.services.BooksService;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -25,19 +25,20 @@ import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 import org.testcontainers.junit.jupiter.Testcontainers;
 
+import java.util.Optional;
+
 @DataJpaTest
 @Transactional(propagation = Propagation.NOT_SUPPORTED)
 @AutoConfigureTestDatabase(replace = AutoConfigureTestDatabase.Replace.NONE)
 @Testcontainers
 @Import({BooksService.class, DbBooksRepository.class})
-class BooksServiceCreateBookTests extends DatabaseSuite {
+class BooksServiceTest extends DatabaseSuite {
   @Autowired private JpaAuthorsRepository jpaAuthorsRepository;
-
   @Autowired private JpaBooksRepository jpaBooksRepository;
-
   @Autowired private BooksService booksService;
-
+  @Autowired private JpaTagsRepository jpaTagsRepository;
   @MockBean private AuthorsRegistryServiceGatewayBase authorsGateway;
+  private Book testBook;
 
   private Author testAuthor;
 
@@ -45,7 +46,12 @@ class BooksServiceCreateBookTests extends DatabaseSuite {
   public void setUp() {
     jpaBooksRepository.deleteAll();
     jpaAuthorsRepository.deleteAll();
-    testAuthor = jpaAuthorsRepository.save(new Author("Test", "Author"));
+
+    testAuthor = new Author("Test", "Author");
+    jpaAuthorsRepository.save(testAuthor);
+
+    testBook = new Book("Test Book", testAuthor);
+    jpaBooksRepository.save(testBook);
   }
 
   @Test
@@ -56,7 +62,7 @@ class BooksServiceCreateBookTests extends DatabaseSuite {
             testAuthor.getId(), "Экстремальное программирование. Разработка через тестирование");
 
     assertDoesNotThrow(() -> booksService.createNew(info));
-    assertEquals(1, jpaBooksRepository.findAll().size());
+    assertEquals(2, jpaBooksRepository.findAll().size());
   }
 
   @Test
@@ -82,7 +88,7 @@ class BooksServiceCreateBookTests extends DatabaseSuite {
             "Экстремальное программирование. Разработка через тестирование");
 
     assertThrows(InvalidBookDataException.class, () -> booksService.createNew(info));
-    assertEquals(0, jpaBooksRepository.findAll().size());
+    assertEquals(1, jpaBooksRepository.findAll().size());
   }
 
   @Test
@@ -93,5 +99,59 @@ class BooksServiceCreateBookTests extends DatabaseSuite {
             testAuthor.getId(), "Экстремальное программирование. Разработка через тестирование");
 
     assertThrows(InvalidBookDataException.class, () -> booksService.createNew(info));
+  }
+
+  @Test
+  public void testSimpleDelete() {
+    when(authorsGateway.isAuthorWroteThisBook(any(), any(), any())).thenReturn(true);
+    assertTrue(booksService.deleteBook(testBook.getId()));
+    assertEquals(0, jpaBooksRepository.findAll().size());
+  }
+
+  @Test
+  public void testDeleteBookNotExists() {
+    boolean isDeleted = booksService.deleteBook(testBook.getId() + 1);
+
+    Assertions.assertFalse(isDeleted);
+  }
+
+  @Test
+  public void testSimpleFind() {
+    var result = booksService.findBook(testBook.getId());
+
+    Assertions.assertTrue(result.isPresent());
+    Assertions.assertEquals(testBook.getId(), result.get().getId());
+  }
+
+  @Test
+  public void updateAuthor() {
+    var anotherAuthor = new Author("Another", "Author");
+    jpaAuthorsRepository.save(anotherAuthor);
+
+    var book = booksService.updateBookAuthor(testBook.getId(), anotherAuthor.getId());
+
+    assertTrue(book.isPresent());
+    assertEquals(testBook.getId(), book.get().getId());
+    assertEquals(anotherAuthor.getId(), book.get().getAuthor().getId());
+  }
+
+  @Test
+  public void updateTitle() {
+    var book = booksService.updateBookTitle(testBook.getId(), "Another Test Book");
+
+    assertTrue(book.isPresent());
+    assertEquals(testBook.getId(), book.get().getId());
+    assertEquals("Another Test Book", book.get().getTitle());
+  }
+
+  @Test
+  public void updateAddTag() {
+    var tag = new Tag("Some Test Tag");
+    jpaTagsRepository.save(tag);
+
+    Optional<Book> book = booksService.addNewTag(testBook.getId(), tag.getId());
+
+    assertTrue(book.isPresent());
+    assertEquals(1, book.get().getTags().size());
   }
 }
